fix warnings on implicit type conversions in various places

+++ include/hypro/algorithms/reachability/workers/LTIWorker.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022.
+ * Copyright (c) 2022-2023.
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
  * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
@@ -43,7 +43,7 @@ class LTIWorker {
 		, mSubspace( subspace ) {}
 
 	template <typename OutputIt>
-	REACHABILITY_RESULT computeTimeSuccessors( const Representation& initialSet, LocationT const* loc, OutputIt out, int segmentsToCompute, bool checkSafety = true ) const;
+	REACHABILITY_RESULT computeTimeSuccessors( const Representation& initialSet, LocationT const* loc, OutputIt out, long segmentsToCompute, bool checkSafety = true ) const;
 
 	template <typename OutputIt>
 	REACHABILITY_RESULT computeTimeSuccessors( const Representation& initialSet, LocationT const* loc, OutputIt out, bool checkSafety = true ) const {
@@ -63,7 +63,7 @@ class LTIWorker {
 	TimeTransformationCache<LocationT>& mTrafoCache;  ///< cache for matrix exponential
 	std::size_t mSubspace;
 
-	size_t const mNumSegments = size_t( std::ceil( std::nextafter( carl::convert<tNumber, double>( mLocalTimeHorizon / mSettings.timeStep ), std::numeric_limits<double>::max() ) ) );
+	long const mNumSegments = long( std::ceil( std::nextafter( carl::convert<tNumber, double>( mLocalTimeHorizon / mSettings.timeStep ), std::numeric_limits<double>::max() ) ) );
 };
 
 }  // namespace hypro
